requirements.txt
requests~=2.28.1
Pillow~=9.2.0
python-slugify~=7.0.0